feat: chip distribution sidebar with FWHM peak boundaries
- Move chip distribution chart to K-line right sidebar (220px flex layout)
- Add chip_peak_upper/chip_peak_lower fields to AssetAnalysisCard contract
- Implement FWHM (Full Width at Half Maximum) algorithm for peak boundary detection
- Add 5 markLines: peak price (solid yellow), upper/lower boundaries (dashed amber), current price, average cost
- Redesign sidebar stats: peak price, upper boundary, lower boundary, avg cost, profit/loss ratios
- Fix merge conflicts with upstream master (CHANGELOG, ARCHITECTURE, wind client, py_file_index)

+++ core/contracts/industry_chain.py
@@ -186,7 +186,7 @@ class PropagationPath(BaseModel):
         product = 1.0
         for step in self.steps:
             product *= step.mapping_strength
-        return product ** (1 / len(self.steps))
+        return float(product ** (1 / len(self.steps)))
 
 
 class ThesisCard(BaseModel):
